+++ nautilus_core/adapters/src/databento/python/historical.rs
@@ -13,9 +13,12 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use std::{fs, num::NonZeroU64, sync::Arc};
+use std::{fs, num::NonZeroU64, str::FromStr, sync::Arc};
 
-use databento::{dbn, historical::timeseries::GetRangeParams};
+use databento::{
+    dbn::{self, SType},
+    historical::timeseries::GetRangeParams,
+};
 use indexmap::IndexMap;
 use nautilus_core::{
     python::to_pyvalue_err,
@@ -36,7 +39,7 @@ use tokio::sync::Mutex;
 
 use super::loader::convert_instrument_to_pyobject;
 use crate::databento::{
-    common::get_date_time_range,
+    common::{check_consistent_symbology, get_date_time_range, infer_symbology_type},
     decode::{
         decode_imbalance_msg, decode_instrument_def_msg, decode_record, decode_statistics_msg,
         raw_ptr_to_ustr,
@@ -110,19 +113,22 @@ impl DatabentoHistoricalClient {
         &self,
         py: Python<'py>,
         dataset: String,
-        symbols: String,
+        symbols: Vec<&str>,
         start: UnixNanos,
         end: Option<UnixNanos>,
         limit: Option<u64>,
     ) -> PyResult<&'py PyAny> {
         let client = self.inner.clone();
 
+        let stype_in = infer_symbology_type(symbols.first().unwrap());
+        check_consistent_symbology(symbols.as_slice()).map_err(to_pyvalue_err)?;
         let end = end.unwrap_or(self.clock.get_time_ns());
         let time_range = get_date_time_range(start, end).map_err(to_pyvalue_err)?;
         let params = GetRangeParams::builder()
             .dataset(dataset)
             .date_time_range(time_range)
             .symbols(symbols)
+            .stype_in(SType::from_str(&stype_in).map_err(to_pyvalue_err)?)
             .schema(dbn::Schema::Definition)
             .limit(limit.and_then(NonZeroU64::new))
             .build();
@@ -175,19 +181,22 @@ impl DatabentoHistoricalClient {
         &self,
         py: Python<'py>,
         dataset: String,
-        symbols: String,
+        symbols: Vec<&str>,
         start: UnixNanos,
         end: Option<UnixNanos>,
         limit: Option<u64>,
     ) -> PyResult<&'py PyAny> {
         let client = self.inner.clone();
 
+        let stype_in = infer_symbology_type(symbols.first().unwrap());
+        check_consistent_symbology(symbols.as_slice()).map_err(to_pyvalue_err)?;
         let end = end.unwrap_or(self.clock.get_time_ns());
         let time_range = get_date_time_range(start, end).map_err(to_pyvalue_err)?;
         let params = GetRangeParams::builder()
             .dataset(dataset)
             .date_time_range(time_range)
             .symbols(symbols)
+            .stype_in(SType::from_str(&stype_in).map_err(to_pyvalue_err)?)
             .schema(dbn::Schema::Mbp1)
             .limit(limit.and_then(NonZeroU64::new))
             .build();
@@ -239,19 +248,22 @@ impl DatabentoHistoricalClient {
         &self,
         py: Python<'py>,
         dataset: String,
-        symbols: String,
+        symbols: Vec<&str>,
         start: UnixNanos,
         end: Option<UnixNanos>,
         limit: Option<u64>,
     ) -> PyResult<&'py PyAny> {
         let client = self.inner.clone();
 
+        let stype_in = infer_symbology_type(symbols.first().unwrap());
+        check_consistent_symbology(symbols.as_slice()).map_err(to_pyvalue_err)?;
         let end = end.unwrap_or(self.clock.get_time_ns());
         let time_range = get_date_time_range(start, end).map_err(to_pyvalue_err)?;
         let params = GetRangeParams::builder()
             .dataset(dataset)
             .date_time_range(time_range)
             .symbols(symbols)
+            .stype_in(SType::from_str(&stype_in).map_err(to_pyvalue_err)?)
             .schema(dbn::Schema::Trades)
             .limit(limit.and_then(NonZeroU64::new))
             .build();
@@ -304,7 +316,7 @@ impl DatabentoHistoricalClient {
         &self,
         py: Python<'py>,
         dataset: String,
-        symbols: String,
+        symbols: Vec<&str>,
         aggregation: BarAggregation,
         start: UnixNanos,
         end: Option<UnixNanos>,
@@ -312,6 +324,8 @@ impl DatabentoHistoricalClient {
     ) -> PyResult<&'py PyAny> {
         let client = self.inner.clone();
 
+        let stype_in = infer_symbology_type(symbols.first().unwrap());
+        check_consistent_symbology(symbols.as_slice()).map_err(to_pyvalue_err)?;
         let schema = match aggregation {
             BarAggregation::Second => dbn::Schema::Ohlcv1S,
             BarAggregation::Minute => dbn::Schema::Ohlcv1M,
@@ -325,6 +339,7 @@ impl DatabentoHistoricalClient {
             .dataset(dataset)
             .date_time_range(time_range)
             .symbols(symbols)
+            .stype_in(SType::from_str(&stype_in).map_err(to_pyvalue_err)?)
             .schema(schema)
             .limit(limit.and_then(NonZeroU64::new))
             .build();
@@ -377,19 +392,22 @@ impl DatabentoHistoricalClient {
         &self,
         py: Python<'py>,
         dataset: String,
-        symbols: String,
+        symbols: Vec<&str>,
         start: UnixNanos,
         end: Option<UnixNanos>,
         limit: Option<u64>,
     ) -> PyResult<&'py PyAny> {
         let client = self.inner.clone();
 
+        let stype_in = infer_symbology_type(symbols.first().unwrap());
+        check_consistent_symbology(symbols.as_slice()).map_err(to_pyvalue_err)?;
         let end = end.unwrap_or(self.clock.get_time_ns());
         let time_range = get_date_time_range(start, end).map_err(to_pyvalue_err)?;
         let params = GetRangeParams::builder()
             .dataset(dataset)
             .date_time_range(time_range)
             .symbols(symbols)
+            .stype_in(SType::from_str(&stype_in).map_err(to_pyvalue_err)?)
             .schema(dbn::Schema::Imbalance)
             .limit(limit.and_then(NonZeroU64::new))
             .build();
@@ -431,19 +449,22 @@ impl DatabentoHistoricalClient {
         &self,
         py: Python<'py>,
         dataset: String,
-        symbols: String,
+        symbols: Vec<&str>,
         start: UnixNanos,
         end: Option<UnixNanos>,
         limit: Option<u64>,
     ) -> PyResult<&'py PyAny> {
         let client = self.inner.clone();
 
+        let stype_in = infer_symbology_type(symbols.first().unwrap());
+        check_consistent_symbology(symbols.as_slice()).map_err(to_pyvalue_err)?;
         let end = end.unwrap_or(self.clock.get_time_ns());
         let time_range = get_date_time_range(start, end).map_err(to_pyvalue_err)?;
         let params = GetRangeParams::builder()
             .dataset(dataset)
             .date_time_range(time_range)
             .symbols(symbols)
+            .stype_in(SType::from_str(&stype_in).map_err(to_pyvalue_err)?)
             .schema(dbn::Schema::Statistics)
             .limit(limit.and_then(NonZeroU64::new))
             .build();